+++ airbyte-integrations/connectors/destination-clickhouse-v2/src/main/kotlin/io/airbyte/integrations/destination/clickhouse_v2/config/ClickhouseBeanFactory.kt
@@ -1,30 +1,15 @@
-package io.airbyte.integrations.destination.clickhouse_v2
+package io.airbyte.integrations.destination.clickhouse_v2.config
 
 import com.clickhouse.client.api.Client
-import com.clickhouse.data.ClickHouseDataType
 import io.airbyte.cdk.command.ConfigurationSpecificationSupplier
-import io.airbyte.cdk.load.client.AirbyteClient
-import io.airbyte.cdk.load.orchestration.db.DatabaseInitialStatusGatherer
-import io.airbyte.cdk.load.orchestration.db.direct_load_table.DirectLoadInitialStatus
 import io.airbyte.integrations.destination.clickhouse_v2.spec.ClickhouseSpecification
 import io.airbyte.integrations.destination.clickhouse_v2.spec.ClickhouseConfiguration
 import io.airbyte.integrations.destination.clickhouse_v2.spec.ClickhouseConfigurationFactory
-import io.airbyte.integrations.destination.clickhouse_v2.write.direct.ClickhouseDirectLoadDatabaseInitialStatusGatherer
 import io.micronaut.context.annotation.Factory
 import jakarta.inject.Singleton
 
 @Factory
 class ClickhouseBeanFactory {
-
-    @Singleton
-    fun tableStatusGatherer(
-        airbyteClient: AirbyteClient,
-        clickhouseConfiguration: ClickhouseConfiguration,
-    ): DatabaseInitialStatusGatherer<DirectLoadInitialStatus> =
-        ClickhouseDirectLoadDatabaseInitialStatusGatherer(
-            airbyteClient,
-            clickhouseConfiguration,
-    )
 
     @Singleton
     fun clickhouseClient(config: ClickhouseConfiguration): Client {
