+++ airbyte-integrations/connectors/destination-clickhouse-v2/src/main/kotlin/io/airbyte/integrations/destination/clickhouse_v2/write/ClickhouseDirectLoadDatabaseInitialStatusGatherer.kt
@@ -1,10 +1,12 @@
-package io.airbyte.integrations.destination.clickhouse_v2.write.direct
+package io.airbyte.integrations.destination.clickhouse_v2.write
 
 import io.airbyte.cdk.load.client.AirbyteClient
 import io.airbyte.cdk.load.orchestration.db.BaseDatabaseInitialStatusGatherer
 import io.airbyte.cdk.load.orchestration.db.direct_load_table.DirectLoadInitialStatus
 import io.airbyte.integrations.destination.clickhouse_v2.spec.ClickhouseConfiguration
+import jakarta.inject.Singleton
 
+@Singleton
 class ClickhouseDirectLoadDatabaseInitialStatusGatherer(
     airbyteClient: AirbyteClient,
     clickhouseConfiguration: ClickhouseConfiguration,
